fix: rename core logger to logging
