+++ packages/core/src/utils/getClientInfo/getClientInfo.ts
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { ConsoleLogger } from '../../logger';
+import { ConsoleLogger } from '../../logging';
 
 const logger = new ConsoleLogger('getClientInfo');
